Remove old contract store functionality.
